Returning a vector of doubles in the logminernumview instead of integers vector

+++ src/edu/nyu/cs/cs2580/LogMinerNumviews.java
@@ -58,7 +58,7 @@ public class LogMinerNumviews extends LogMiner {
     //Make the index directory if not exists
     File indexDir = new File(_options._indexPrefix);
     indexDir.mkdir();
-    
+
     File newFile = new File(_options._indexPrefix+ "/numViewsIndex.idx");
 
 
@@ -144,9 +144,9 @@ public class LogMinerNumviews extends LogMiner {
     System.out.println("Loading using " + this.getClass().getName());
     Scanner sc = new Scanner(new File(_options._indexPrefix+ "/numViewsIndex.idx"));
 
-    Vector<Integer> numViewIndex = new Vector<Integer>();
+    Vector<Double> numViewIndex = new Vector<Double>();
     while (sc.hasNext()){
-      numViewIndex.add(Integer.parseInt(sc.next()));
+      numViewIndex.add((double)Integer.parseInt(sc.next()));
 
     }
